Moved DATE_FIELD_VALUE_TO_LOCALDATETIME to LineFileDocs, added a test case to TestLineFileDocs
