+++ lucene/test-framework/src/test/org/apache/lucene/tests/util/TestLineFileDocs.java
@@ -1,0 +1,36 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.lucene.tests.util;
+
+import java.time.LocalDateTime;
+
+public class TestLineFileDocs extends LuceneTestCase {
+  /**
+   * Tests that {@link #expectThrows} behaves correctly when the Runnable throws (an instance of a
+   * subclass of) the expected Exception type: by returning that Exception.
+   */
+  public void testDateFieldNormalization() {
+    // europarl corpus uses this data format.
+    assertEquals(
+        LocalDateTime.of(2023, 2, 23, 0, 0),
+        LineFileDocs.DATE_FIELD_VALUE_TO_LOCALDATETIME.apply("2023-02-23"));
+    // enwiki uses this data format.
+    assertEquals(
+        LocalDateTime.of(2010, 1, 12, 12, 32, 45),
+        LineFileDocs.DATE_FIELD_VALUE_TO_LOCALDATETIME.apply("12-JAN-2010 12:32:45.000"));
+  }
+}
